articles_fetcher/requirements.txt
firebase-admin==5.2.0
google-cloud-tasks==2.8.0
google-auth==2.6.0